intentando hacer que funcione todo

+++ agenda.php
@@ -29,9 +29,6 @@
 		<div id="datos">
 			<h3>Agenda</h3>
 				<?php
-					$nombre=strtolower($_POST["nombre"]);
-					$correo=strtolower($_POST["correo"]);
-
 					if(!isset($_SESSION["agenda"])){
 						$_SESSION["agenda"]= array();
 					}
@@ -39,65 +36,33 @@
 					{
 						$agenda=$_SESSION["agenda"];
 						if($_POST["guardar"]){
-							$agenda[$nombre]=$correo;
-							//$agenda[limpiar_String($nombre)]=limpiar_String($correo);
-							foreach ($agenda as $key => $value){
-								echo "Nombre: ".$key."<br>";
-								echo "Correo: ".$value."<br>";
-								echo "<hr>";
+							$nombre=strtolower($_POST["nombre"]);
+							$correo=strtolower($_POST["correo"]);
+							if(array_key_exists($nombre,$agenda)){
+								if($correo==""){
+									unset($agenda[$nombre]);
+								}
+								else
+								{
+									$agenda[$nombre]=$correo;
+								}				
+							}elseif($nombre="" && $email=""){
+								echo "inserte datos en los campos por favor.";
+							}
+							else{
+								$agenda[$nombre]=$correo;
+								foreach ($agenda as $key => $value){
+									echo "Nombre: ".$key."<br>";
+									echo "Correo: ".$value."<br>";
+									echo "<hr>";
+								}
+
 							}
 						}
 						$_SESSION["agenda"]=$agenda;
 					}
 
-					function limpiar_String($string){
- 
-    					$string = trim($string);
- 
-    					$string = str_replace(
-        					array('á', 'à', 'ä', 'â', 'ª', 'Á', 'À', 'Â', 'Ä'),
-        					array('a', 'a', 'a', 'a', 'a', 'A', 'A', 'A', 'A'),
-        					$string
-        				);
- 						
-					    $string = str_replace(
-					        array('é', 'è', 'ë', 'ê', 'É', 'È', 'Ê', 'Ë'),
-					        array('e', 'e', 'e', 'e', 'E', 'E', 'E', 'E'),
-					     	$string
-					    );
-					 
-					    $string = str_replace(
-					        array('í', 'ì', 'ï', 'î', 'Í', 'Ì', 'Ï', 'Î'),
-					        array('i', 'i', 'i', 'i', 'I', 'I', 'I', 'I'),
-					        $string
-					    );
- 
-					    $string = str_replace(
-					        array('ó', 'ò', 'ö', 'ô', 'Ó', 'Ò', 'Ö', 'Ô'),
-					        array('o', 'o', 'o', 'o', 'O', 'O', 'O', 'O'),
-					        $string
-					    );
- 
-					    $string = str_replace(
-					        array('ú', 'ù', 'ü', 'û', 'Ú', 'Ù', 'Û', 'Ü'),
-					        array('u', 'u', 'u', 'u', 'U', 'U', 'U', 'U'),
-					        $string
-					    ); 
-    					$string = str_replace(
-					        array("\", "¨", "º", "-", "~",
-					             "#", "@", "|", "!", """,
-					             "·", "$", "%", "&", "/",
-					             "(", ")", "?", "'", "¡",
-					             "¿", "[", "^", "<code>", "]",
-					             "+", "}", "{", "¨", "´",
-					             ">", "< ", ";", ",", ":",
-					             ".", " "),
-					        '',
-					        $string
-    					);
- 						
- 						return $string;
- 					}
+					
 					?>
 		</div>
 	</body>
